+++ OpenOrchestrator/database/truncated_string.py
@@ -1,8 +1,10 @@
 """Custom sqlalchemy Type for generating limited String objects"""
 
 import math
+from typing import Any
 
-from sqlalchemy import String, TypeDecorator
+from sqlalchemy import Dialect
+from sqlalchemy.types import String, TypeDecorator
 
 
 class LimitedLengthString(TypeDecorator):
@@ -11,6 +13,16 @@ class LimitedLengthString(TypeDecorator):
 
     def process_bind_param(self, value, dialect):
         return _truncate_message(value)
+
+    def process_literal_param(self, value: Any | None, dialect: Dialect) -> str:
+        return super().process_literal_param(value, dialect)
+
+    def process_result_value(self, value: Any | None, dialect: Dialect) -> Any | None:
+        return super().process_result_value(value, dialect)
+
+    @property
+    def python_type(self) -> String:
+        return String
 
 
 def _truncate_message(message: str, max_length: int = 8000) -> str: